Refactor function name from 'makeitshort' to 'request_api'
In the previous version, the 'makeitshort' function name was slightly misleading as it implied that the only purpose of this method was to shorten URLs. However, the function is also responsible for making requests to an API. With the change, 'request_api' provides a better understanding and readability, accurately describing its functionality.