Refactor API related code and provide URL default value
The code related to API handling was refactored, transitioning to GET requests instead of POST for an updated API endpoint, and adjusting the response struct to suit the changes. Also, a default value was provided for the URL argument to enhance usability whenever a URL is not specified by the user.